requirements.txt
flask
yt-dlp
pywebview 